move water assessment to its own files
