+++ tests/wasm/wasm-express.js
@@ -14,7 +14,8 @@ app.get('/', function(req, res) {
     res.sendFile(path.join(__dirname + '/wasm-express.html'));
 });
 
-app.use('/assets', express.static('bin'));
+app.use('/tests', express.static('tests/wasm/electrical/'));
+app.use('/tests', express.static('tests/wasm/water/'));
 app.use('/tests', express.static('tests/wasm/test-helpers/'));
 app.use('/tests', express.static('tests/wasm/psat/'));
 app.use('/tests', express.static('tests/wasm/motor/'));
